mini_project_2/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
pymongo==4.6.1
mysqlclient==2.2.1 
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
PyPDF2==3.0.1
python-docx==1.0.1
spacy==3.7.2
nltk==3.8.1
drf-yasg==1.21.7